requirements.txt
aiogram==3.22.0
python-dotenv==1.1.1
requests==2.32.5
PyPDF2==3.0.1
langchain==0.3.27
openai==2.4.0

